+++ src/main/java/br/ufscar/dc/dsw/BugBountyPlatform/service/impl/ProgramaService.java
@@ -1,10 +1,11 @@
 package br.ufscar.dc.dsw.BugBountyPlatform.service.impl;
 
 import br.ufscar.dc.dsw.BugBountyPlatform.dao.IProgramaDAO;
+import br.ufscar.dc.dsw.BugBountyPlatform.dao.IRelatorioDAO;
+import br.ufscar.dc.dsw.BugBountyPlatform.domain.Empresa;
 import br.ufscar.dc.dsw.BugBountyPlatform.domain.Programa;
 import br.ufscar.dc.dsw.BugBountyPlatform.service.IProgramaService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -16,6 +17,9 @@ public class ProgramaService implements IProgramaService {
 
     @Autowired
     private IProgramaDAO programaDAO;
+
+    @Autowired
+    private IRelatorioDAO relatorioDAO;
 
     @Override
     @Transactional(readOnly = true)
@@ -36,32 +40,36 @@ public class ProgramaService implements IProgramaService {
 
     @Override
     public boolean excluir(Long id) {
-        try {
-            programaDAO.deleteById(id);
-            return true;
-        }
-        catch (DataIntegrityViolationException e) {
+        Programa programa = this.buscarPorId(id);
+        if (programa == null) return false;
+
+        if (!relatorioDAO.findByPrograma(programa).isEmpty())
             return false;
-        }
+
+        programaDAO.deleteById(id);
+        return true;
     }
 
     @Override
-    public List<Programa> buscarPorSetor(String setor) {
-        return programaDAO.findByEmpresaSetorContainingIgnoreCase(setor);
+    public List<Programa> buscarPorFiltro(String termo) {
+        return programaDAO.findByEmpresaNomeContainingIgnoreCaseOrEmpresaSetorContainingIgnoreCase(termo, termo);
     }
 
     @Override
     public void atualizar(Programa programaForm) {
         Programa programaBanco = this.buscarPorId(programaForm.getId());
-
         if (programaBanco != null) {
             programaBanco.setTitulo(programaForm.getTitulo());
             programaBanco.setEscopo(programaForm.getEscopo());
             programaBanco.setRecompensaMaxima(programaForm.getRecompensaMaxima());
             programaBanco.setDataLimite(programaForm.getDataLimite());
-            programaBanco.setEmpresa(programaForm.getEmpresa()); // Caso a edição permita trocar a empresa
-
+            programaBanco.setEmpresa(programaForm.getEmpresa());
             programaDAO.save(programaBanco);
         }
+    }
+
+    @Override
+    public List<Programa> buscarPorEmpresa(Empresa empresa) {
+        return programaDAO.findByEmpresa(empresa);
     }
 }
